# docs/openapi.platform.yaml
openapi: 3.1.0
info:
  title: AI Control Tower Platform API
  version: 1.0.0
  description: >
    Broader platform OpenAPI reference for AI Control Tower covering authentication,
    enterprise identity, tenant admin, registry, controls, workflows, audit, notifications,
    evidence, exports, risk, dashboard, and marketing capture endpoints.
  license:
    name: Proprietary
servers:
  - url: https://enterprise-ai-governance.onrender.com
    description: Production backend
  - url: http://localhost:5000
    description: Local development
...
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
...
  schemas:
...
    LeadCapture:
      type: object
      properties:
        name:
          type: string
        workEmail:
          type: string
          format: email
        company:
          type: string
        role:
          type: string
        teamSize:
          type: string
        primaryChallenge:
          type: string
        formType:
          type: string
          enum: [book_demo, start_pilot]
        source:
          type: [string, 'null']
        ctaSource:
          type: [string, 'null']
        campaign:
          type: [string, 'null']
        notes:
          type: [string, 'null']
      required: [name, workEmail, company, role, teamSize, primaryChallenge, formType]
    MarketingEvent:
      type: object
      properties:
        eventName:
          type: string
        pagePath:
          type: [string, 'null']
        section:
          type: [string, 'null']
        cta:
          type: [string, 'null']
        source:
          type: [string, 'null']
        campaign:
          type: [string, 'null']
        referrer:
          type: [string, 'null']
        metadata:
          type: object
          additionalProperties: true
      required: [eventName]
...
paths:
...
  /api/track:
    post:
      tags: [Marketing]
      summary: Record a marketing event
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketingEvent'
      responses:
        '200':
          description: Event captured
  /api/leads:
    post:
      tags: [Marketing]
      summary: Capture demo or pilot lead
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadCapture'
      responses:
        '201':
          description: Lead captured
    get:
      tags: [Marketing]
      summary: List captured leads
      responses:
        '200':
          description: Lead list
...